Move TCK tests into issues file
